requirements.txt
pytest==5.3.2
polling
PyHamcrest
pytest-shell